+++ src/test/auxiliary/crateresolve1-2.rs
@@ -12,4 +12,4 @@
 #![crate_name = "crateresolve1"]
 #![crate_type = "lib"]
 
-pub fn f() -> int { 20 }
+pub fn f() -> isize { 20 }
